refactors support page to split into separate components:
